requirements.txt
streamlit
opencv-python
numpy
ultralytics
pillow
onvif-zeep